requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
numpy==1.21.1
panda==0.3.1
pandas==1.3.1
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
spotipy==2.18.0
toml==0.10.2
urllib3==1.26.6
